+++ engine/examples/multiLanguageTest/game.js
@@ -3,7 +3,7 @@ import RPGinia from "/src/RPGinia.js";
 const engine = new RPGinia();
 const app = new engine.App("Test RPGinia app");
 const load = new app.Loaders();
-const world = new app.World();
+const world = new app.World(true);
 const kb = new app.Keyboard();
 const cam = new app.Camera();
 const audio = new app.AudioManager();