+++ project-editor/src/renderer/assets/js/createNewProjectFormValidation.js
@@ -1,3 +1,0 @@
-export default function validate(name, type, dir) {
-    console.log("ti pidor", name, type, dir)
-}