+++ project-editor/src/renderer/components/EditorComponents/ObjectProperties.vue
@@ -1,0 +1,13 @@
+<template>
+    <Rectangle :object="object.settings"></Rectangle>
+</template>
+<script>
+import Rectangle from './Objects/Rectangle';
+
+export default {
+    components: { Rectangle },
+    props: {
+        object: Object
+    }
+}
+</script>