+++ project-editor/src/renderer/components/ProjectExistingError.vue
@@ -7,15 +7,12 @@
             <div class="description">
                 <p>
                     You can 
-                    <a href="#" id="createProject" @click="createModalOpened = true">create new</a> 
+                    <a href="#" id="createProject" @click="createProject">create new</a> 
                     or 
                     <a href="#" id="openProject">open existing</a>
                 </p>
             </div>
         </div>
-        <Modal title="Create new project" v-if="createModalOpened" @close="createModalOpened = false">
-            <NewProjectModalContent></NewProjectModalContent>
-        </Modal>
     </div>
 </template>
 <script>
@@ -27,16 +24,12 @@ import NewProjectModalContent from './Modals/NewProject';
 export default {
     name: 'existing-error',
     components: { Modal, NewProjectModalContent },
-    data: function() {
-        return {
-            createModalOpened: false
+    methods: {
+        createProject: function() {
+            ipcRenderer.send('createProject')
         }
     }
 }
 </script>
 
 <style src="../assets/css/projectExistingError.css"></style>
-<style src="../assets/css/main.css"></style>
-<style src="../assets/css/fonts.css"></style>
-<style src="../assets/css/themes/dark.css"></style>
-<style src="../assets/css/scrollbar.css"></style>