+++ project-editor/src/renderer/main.js
@@ -5,9 +5,15 @@ import App from './App'
 import router from './router'
 import store from './store'
 
+import { ipcRenderer } from 'electron'
+
 if (!process.env.IS_WEB) Vue.use(require('vue-electron'))
 Vue.http = Vue.prototype.$http = axios
 Vue.config.productionTip = false
+
+ipcRenderer.on('projectNotExist', e => {
+  router.push('existing-error');
+});
 
 /* eslint-disable no-new */
 new Vue({