+++ CHANGELOG.md
@@ -1,8 +1,17 @@
 # RPGinia change log
 
-## v0.1.0
-* Made first release.
+## v0.1.2 (Not published yet)
+### Editor changes:
+* The packager to `.exe` was changed from electron-packager to electron-builder
+
+### Engine changes:
+* `Camera` class have his own coordination system.
+* Now you can set up new canvas sizes and get them.
 
 ## v0.1.1
 * English text has been added to README.
 * Added project editor. It's not ready and was made only a small part of GUI. Not recommended for use.
+* Added `tests` directory in directory `engine`
+
+## v0.1.0
+* Made first release.